feat(user): 完善用户信息与播放历史功能
- 新增用户关注、粉丝列表接口
- 实现用户播放历史记录的增删改查逻辑
- 添加密码加密与校验工具类，提升安全性
- 扩展用户实体性别字段，完善用户信息
- 引入 spring-security-crypto 依赖支持密码加密
- 优化歌单列表查询及数据转换逻辑
- 增加播放次数统计并更新数据库
- 提供用户信息获取与更新接口
- 修复部分实体字段类型不匹配问题

+++ src/main/java/com/yi/musiclisten/entity/Playlist.java
@@ -54,7 +54,7 @@ public class Playlist implements Serializable {
      * 歌单类型：1普通 2收藏歌单
      */
     @TableField("type")
-    private Byte type;
+    private Integer type;
 
     /**
      * 是否公开
